Add files via upload
Write a simple calculator program that repeatedly allows the user to select +,-,*,/ from a menu of operations on real values and then enter the two operands which then computes the result of applying the selected operation to those operands. However, the only named variables you may use are pointer variables; all others must be anonymous variables.